Update 3
Added transpose function that takes a 2rowx2col matrix and transposes the original row to a new column and the original column to a new row.  Updated the multiply function to  perform scalar multiplication on two object matrices and return the result in a new matrix of values.  Updated the sketch to show the tables on the console.