Update 4
Changed code array to data instead of matrix for readability.  Added static versions of the main matrix functions.  Added map function that maps a function to each value in the data array.  Added weights to the input to hidden connections and the hidden to output connections.  Added sigmoud function to calculate exponents.  Added nn calls to the sketch to test.