Update 6
Added a delta sigmoid function that calculates the delta sigmoid of an input.  Added a learning rate to the neural network.  Added gradient mapping to the outputs and errors.  Added delta calculation functionality to the hidden transposed matrix and the hidden to output gradients.  Added functionality that adds the new calculated weights and gradients to the hidden to output waits.  Added bias to the output based on the gradients.  Added functionality to calculate delta change in inputs to hidden section of the network.  Added the bias to the hidden gradient.  Added sample training data and a loop to test the network on the training data.  Mostly working.